requirements.txt
# Human Action Classification - KTH pipeline
# Python 3.8+

# Video I/O
opencv-python>=4.8.0

# Numerics & ML
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Data & config
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.66.0
kagglehub>=0.2.0

# Plotting (for confusion matrices)
matplotlib>=3.7.0
seaborn>=0.12.0
